refactor: PreferenceConfig 拆分组件
